+++ app/src/services/application/index.js
@@ -1,2 +1,2 @@
-export * from './Application.js';
+export * from './CommiteeApplication.js';
 export * from './application.service.js';
